+++ src/plonemeeting/portal/core/utils.py
@@ -54,7 +54,7 @@ def get_api_url_for_meetings(institution, meeting_UID=None):
         institution.meeting_config_id,
     )
     if meeting_UID:
-        url = "{0}&UID={1}&fullobjects=True".format(url, meeting_UID)
+        url = "{0}&UID={1}&fullobjects=True&b_size=9999".format(url, meeting_UID)
     else:
         url = "{0}{1}".format(url, institution.additional_meeting_query_string_for_list)
     return url
@@ -63,8 +63,8 @@ def get_api_url_for_meetings(institution, meeting_UID=None):
 def get_api_url_for_meeting_items(institution, meeting_UID):
     if not institution.plonemeeting_url or not institution.meeting_config_id:
         return
-    url = "{0}/{1}?sort_on=getItemNumber&privacy=public&privacy=public_heading&" \
-        "getConfigId={2}&linkedMeetingUID={3}&fullobjects=True{4}".format(
+    url = "{0}/{1}?sort_on=getItemNumber&privacy=public&privacy=public_heading&b_size=9999" \
+        "&getConfigId={2}&linkedMeetingUID={3}&fullobjects=True{4}".format(
             institution.plonemeeting_url.rstrip("/"),
             PLONEMEETING_API_MEETING_ITEMS_VIEW,
             institution.meeting_config_id,